Move new configuration class to source root
