+++ application/src/Configuration.test.ts
@@ -1,6 +1,6 @@
 import {access, readFile} from 'fs/promises';
-import Configuration, {ConfigurationLoadError} from './Configuration';
-import {TerseConfiguration, VerboseConfiguration} from '../tests/fixtures/Configuration';
+import {Configuration, ConfigurationLoadError} from './Configuration';
+import {JSONConfiguration, YAMLConfiguration} from '../tests/fixtures/Configuration';
 
 jest.mock('fs/promises', () => ({
   access: jest.fn(),
@@ -9,34 +9,85 @@ jest.mock('fs/promises', () => ({
 
 describe('Configuration', () => {
   describe('Configuration.load', () => {
-    describe.each([
-      ['a valid verbose', VerboseConfiguration],
-      ['a valid terse', TerseConfiguration],
-    ])('%s configuration', (_type: string, configurationInput: ConfigurationFile) => {
+    describe('JSON configuration file', () => {
       let configuration: Configuration;
 
       beforeAll(async () => {
-        (readFile as jest.Mock).mockResolvedValueOnce(Buffer.from(JSON.stringify(configurationInput)));
-        (access as jest.Mock).mockResolvedValueOnce(null);
+        (access as jest.Mock).mockImplementation(async file => {
+          if (file.endsWith('.json')) return null;
+          throw new Error('file does not exist');
+        });
+
+        (readFile as jest.Mock).mockResolvedValueOnce(Buffer.from(JSONConfiguration));
+
         configuration = await Configuration.load('/test');
       });
 
-      test('checks if the configuration file exists', () => {
-        expect(access).toHaveBeenCalledWith('/test/continuous-security.json');
+      afterAll(() => {
+        (access as jest.Mock).mockReset();
       });
 
-      test('attempts to load the configuration file', () => {
-        expect(readFile).toHaveBeenCalledWith('/test/continuous-security.json');
+      test('attempts to read from .continuous-security.json', () => {
+        expect(access).toHaveBeenCalledWith('/test/.continuous-security.json');
+        expect(readFile).toHaveBeenCalledWith('/test/.continuous-security.json');
       });
 
-      test('returns an instantiated Configuration class', () => {
-        expect(configuration).toHaveProperty('scanners', [{name: 'test-scanner'}]);
+      test('loads a list of scanners', () => {
+        expect(configuration).toHaveProperty('scanners', [
+          {name: 'test-scanner'},
+        ]);
+      });
+    });
+    describe('YAML configuration file', () => {
+      let configuration: Configuration;
+
+      beforeAll(async () => {
+        (access as jest.Mock).mockImplementation(async file => {
+          if (file.endsWith('.yaml')) return null;
+          throw new Error('file does not exist');
+        });
+
+        (readFile as jest.Mock).mockResolvedValueOnce(Buffer.from(YAMLConfiguration));
+
+        configuration = await Configuration.load('/test');
+      });
+
+      afterAll(() => {
+        (access as jest.Mock).mockReset();
+      });
+
+      test('attempts to read from .continuous-security.yaml', () => {
+        expect(access).toHaveBeenCalledWith('/test/.continuous-security.yaml');
+        expect(readFile).toHaveBeenCalledWith('/test/.continuous-security.yaml');
+      });
+
+      test('loads a list of scanners', () => {
+        expect(configuration).toHaveProperty('scanners', [
+          {name: 'test-scanner'},
+        ]);
       });
     });
 
     describe('a non-existent configuration file', () => {
-      beforeAll(() => {
-        (access as jest.Mock).mockRejectedValueOnce(new Error('File does not exist'));
+      beforeAll(async () => {
+        (access as jest.Mock).mockRejectedValue(new Error('File does not exist'));
+        await Configuration.load('/test').catch(() => null);
+      });
+
+      afterAll(() => {
+        (access as jest.Mock).mockReset();
+      });
+
+      test('attempts to read from /test/.continuous-security.json', () => {
+        expect(access).toHaveBeenCalledWith('/test/.continuous-security.json');
+      });
+
+      test('attempts to read from /test/.continuous-security.yaml', () => {
+        expect(access).toHaveBeenCalledWith('/test/.continuous-security.yaml');
+      });
+
+      test('attempts to read from /test/.continuous-security.yml', () => {
+        expect(access).toHaveBeenCalledWith('/test/.continuous-security.yml');
       });
 
       test('throws a ConfigurationLoadError', async () => {
